models/geolocation/requirements.txt
openai
pandas
pillow
openpyxl
streamlit
dotenv
google-generativeai
datasets